feat: print config using --print-config
refactor: move command constructing to separate package
